⚡️ :: 중첩 클래스로 처리
