--- data/src/main/java/com/semicolon/data/remote/response/users/userinquiryprofile/TitleBadge.kt
@@ -1,9 +1,0 @@
-package com.semicolon.data.remote.response.users.userinquiryprofile
-
-import com.google.gson.annotations.SerializedName
-
-data class TitleBadge(
-    @SerializedName("id") val id: Int,
-    @SerializedName("image") val image: String,
-    @SerializedName("name") val name: String
-)
